Update TokenAuthentication project to .NET 6.0
Cleaned up and updated the TokenAuthentication project to use .NET 6.0. It included updates to csproj files to utilize newer project format and assembly info and package files were removed as they are no longer needed. Code in Program.cs was made more concise. The SDK version was revised accordingly in the ConvertApi solution file. The goal was to simplify the codebase and adopt up-to-date practices, leading to enhanced maintainability.